requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
distlib==0.3.8
Flask==2.3.3
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.6.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
looseversion==1.3.0
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
mysqlclient==2.2.4
pydantic==2.5.1
pydantic_core==2.14.3
PyJWT==2.8.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.31
urllib3==2.0.5
Werkzeug==2.3.7
python-dotenv==1.0.1
